cleanup files; 2dmap test separate Makefile
