--- .github/workflows/norm-check.yml
@@ -1,22 +1,0 @@
-name: Norminette check
-
-on:
-  push:
-    branches: [ "master" ]
-  pull_request:
-    branches: [ "master" ]
-
-jobs:
-  Compilation:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v3
-    - name: Install norminette
-      run: python3 -m pip install --upgrade pip setuptools
-    - run: python3 -m pip install norminette
-    - run: echo "Norminette installed Successfully"
-    - name: Norminette Check
-      run: norminette ./include ./src
-    - run: echo "Norm Check complete"
