+++ src/2DMap/Makefile
@@ -6,14 +6,14 @@ ifeq ($(OS_NAME), Linux)
     FRAMEWORK           :=
     LINUX_LIBS          := -lXext -lX11
     LINUX_INCLUDES      := -I/usr/include
-    MLX_DIR             := ./minilibx/minilibx-linux
+    MLX_DIR             := ../../minilibx/minilibx-linux
     OS_FLAG             := -D LINUX
 else
     FSANITIZE           := #-fsanitize=address
     FRAMEWORK           := -framework OpenGL -framework AppKit
     LINUX_LIBS          :=
     LINUX_INCLUDES      :=
-    MLX_DIR             := ./minilibx/mlx
+    MLX_DIR             := ../../minilibx/mlx
     OS_FLAG             := -D OSX
 endif
 
@@ -31,18 +31,18 @@ CC                  	:= gcc
 CFLAGS              	:= -Wall -Werror -Wextra $(FSANITIZE) $(OS_FLAG)
 
 OBJ_DIR             	:= ./objects/
-INCLUDE_DIR         	:= ./include/
-SRCS_DIR				:= ./src/
+INCLUDE_DIR         	:= ../../include/
+SRCS_DIR				:= ./
 
 ### minilibx
 MLX			           	:= $(MLX_DIR)/libmlx.a
 
 ### Get_next_line
-GNL_DIR             	:= src/get_next_line
+GNL_DIR             	:= ../get_next_line
 GNL                 	:= $(GNL_DIR)/libgnl.a
 
 ### Libft
-LIBFT_DIR           	:= src/Libft
+LIBFT_DIR           	:= ../Libft
 LIBFT               	:= $(LIBFT_DIR)/libft.a
 
 LIBRARIES           	:= -lmlx -lm -L. -L$(LIBFT_DIR) -lft -L$(GNL_DIR) -lgnl \
@@ -52,14 +52,23 @@ INCLUDES            	:= -I$(LIBFT_DIR) -I$(GNL_DIR) -I$(INCLUDE_DIR) -I$(MLX_DIR
 
 HEADER					:= cub3d.h\
 							enums.h\
-							struct.h
+							struct.h\
+							map2dtest.h
 HEADER_FILES			:= $(addprefix $(INCLUDE_DIR), $(HEADER))
 
-FILES               	:= 
-
+FILES               	:= draw_2Dmap.c\
+							draw_line.c\
+							draw_maptiles.c\
+							pixel_color.c\
+							main.c\
+							../mlx_draw.c\
+							../mlx_init.c
 
 OBJ_LIST                := $(patsubst %.c,%.o,$(FILES))
 OBJS                    := $(addprefix $(OBJ_DIR),$(OBJ_LIST))
+
+SRC_OBJ_LIST			:= $(patsubst %.c,%.o,$(SRC_FILES))
+SRC_OBJS                := $(addprefix $(OBJ_DIR), $(SRC_OBJ_LIST))
 
 ### RULES ###
 all: $(NAME)
@@ -78,7 +87,7 @@ $(NAME): $(OBJ_DIR) $(OBJS) $(HEADER_FILES)
 	@echo "get_next_line done"
 	@make -C ${MLX_DIR}
 	@echo "Compiling..."
-	@$(CC) $(CFLAGS) $(OBJS) $(INCLUDES) $(LIBRARIES) -o $(NAME)
+	@$(CC) $(CFLAGS) $(OBJS) $(SRC_OBJS) $(INCLUDES) $(LIBRARIES) -o $(NAME)
 	@echo "Done."
