+++ src/2DMap/main.c
@@ -1,0 +1,50 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   main.c                                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jdagoy <jdagoy@student.s19.be>             +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/12/17 22:33:41 by jdagoy            #+#    #+#             */
+/*   Updated: 2023/12/19 00:48:08 by jdagoy           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "cub3d.h"
+#include "map2dtest.h"
+
+void	init_testmap(t_gametest *game)
+{
+	int	map[ROWS][COLS] = {
+	{1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
+	{1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1},
+	{1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1},
+	{1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1},
+	{1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1},
+	{1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1},
+	{1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
+	{1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
+	{1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1},
+	{1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
+	{1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1}
+	};
+	ft_memcpy(game->map, map, sizeof(int) * ROWS * COLS);
+	game->map_width = COLS;
+	game->map_height = ROWS;
+}
+
+int	main(void)
+{
+	t_gametest		game;
+	t_display		mlx;
+
+	init_testmap(&game);
+	mlx.mlx = mlx_init();
+	mlx.window = mlx_new_window(mlx.mlx, WIDTH, HEIGHT, "cub3D");
+	mlx.img.img = mlx_new_image(mlx.mlx, WIDTH, HEIGHT);
+	mlx.img.address = mlx_get_data_addr(mlx.img.img,
+			&mlx.img.bits_per_pixel, &mlx.img.line_length, &mlx.img.endian);
+	draw_map(&game, &mlx);
+	mlx_loop(mlx.mlx);
+	return (0);
+}
