+++ src/2DMap/pixel_color.c
@@ -1,13 +1,29 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   draw_line_dda.c                                    :+:      :+:    :+:   */
+/*   pixel_color.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jdagoy <jdagoy@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/12/18 15:01:56 by jdagoy            #+#    #+#             */
-/*   Updated: 2023/12/18 15:02:09 by jdagoy           ###   ########.fr       */
+/*   Created: 2023/12/18 15:19:55 by jdagoy            #+#    #+#             */
+/*   Updated: 2023/12/19 00:57:06 by jdagoy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "cub3d.h"
+void	set_color(char *buffer, int endian, int color, int alpha)
+{
+	if (endian == 1)
+	{
+		buffer[0] = alpha;
+		buffer[1] = (color >> 16) & 0xFF;
+		buffer[2] = (color >> 8) & 0xFF;
+		buffer[3] = (color) & 0xFF;
+	}
+	else
+	{
+		buffer[0] = (color) & 0xFF;
+		buffer[1] = (color >> 8) & 0xFF;
+		buffer[2] = (color >> 16) & 0xFF;
+		buffer[3] = alpha;
+	}
+}
